+++ account_reconcile_model_oca/models/__init__.py
@@ -1,2 +1,3 @@
 from . import account_reconcile_model
+from . import account_reconcile_model_line
 from . import account_bank_statement_line
